feat: new methods for players and coach

+++ ProjetoFinalBackend/src/main/java/com/springboot/projetofinalbackend/service/TrainingService.java
@@ -11,6 +11,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.List;
 import java.util.Set;
@@ -23,7 +26,7 @@ public class TrainingService {
     @Autowired
     private PlayerRepository playerRepository;
 
-    public ResponseEntity<Training> save(TrainingDTO trainingDto){
+    public ResponseEntity<Training> create(@RequestBody TrainingDTO trainingDto){
         var training = new Training();
         BeanUtils.copyProperties(trainingDto, training);
         return ResponseEntity.status(HttpStatus.CREATED).body(trainingRepository.save(training));
@@ -37,29 +40,36 @@ public class TrainingService {
         return ResponseEntity.status(HttpStatus.OK).body(trainingRepository.findById(id).orElse(null));
     }
 
-    public ResponseEntity<Training> update(Training training){
-        var trainingUpdate = trainingRepository.findById(training.getId()).orElse(null);
+    public ResponseEntity<Training> update(@PathVariable Long id ,@RequestBody TrainingDTO training){
+        var trainingUpdate = trainingRepository.findById(id).orElse(null);
         if(trainingUpdate == null){
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
         }
-        training.setTitle(training.getTitle());
-        training.setLocation(training.getLocation());
-        training.setDateTime(training.getDateTime());
-        return ResponseEntity.status(HttpStatus.OK).body(trainingRepository.save(training));
+        trainingUpdate.setTitle(training.title());
+        trainingUpdate.setLocation(training.location());
+        trainingUpdate.setDateTime(training.dateTime());
+        return ResponseEntity.status(HttpStatus.OK).body(trainingRepository.save(trainingUpdate));
     }
 
-    public ResponseEntity<Void> delete(Training training){
-        var trainingDelete = trainingRepository.findById(training.getId()).orElse(null);
+    public ResponseEntity<Void> delete(@PathVariable Long id, @RequestParam String confirmation){
+        var trainingDelete = trainingRepository.findById(id).orElse(null);
         if(trainingDelete == null){
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
+
+        if(!trainingDelete.getTitle().equals(confirmation)){
+            return ResponseEntity.status(HttpStatus.NOT_MODIFIED).build();
+        }
+
         trainingRepository.delete(trainingDelete);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 
     public Set<Player> addPlayer(Long trainingId, Long playerId) {
-        var training = trainingRepository.findById(trainingId).orElseThrow(() -> new EntityNotFoundException("Training not found"));
-        var player = playerRepository.findById(playerId).orElseThrow(() -> new EntityNotFoundException("Player not found"));
+        var training = trainingRepository.findById(trainingId)
+                .orElseThrow(() -> new EntityNotFoundException("Training not found"));
+        var player = playerRepository.findById(playerId)
+                .orElseThrow(() -> new EntityNotFoundException("Player not found"));
 
         if (!training.getPlayers().contains(player)) {
             training.getPlayers().add(player);
@@ -68,6 +78,7 @@ public class TrainingService {
 
         return training.getPlayers();
     }
+
 
     public ResponseEntity<Void> deletePlayer(Long trainingId, Long playerId) {
         var training = trainingRepository.findById(trainingId)
